initial version of Mistral
Initial implementation of the Mistral part of Tesora-dbaas plugin.

No additional Fuel UI configuration exists for the Mistral components.
Mistral services will run on every tesora-dbaas node, alongside trove services.

Mistral database, Keystone, and Haproxy/VIPs are set up very similarly as
we do for Trove.   Controllers load balance accesses to the Mistral API.
Mistral uses the rabbit server that runs on tesora-dbaas nodes, and not 
the one that exists on the controllers.

Change-Id: If5326b9a6f87835922bfaa661c0cea1c469d7ff9