lower timeout for haproxy determining mistral-api is alive
We've seen;
- when you hotplug add dbaas, then deploy the controller only,
you will always hit the 10 minute delay
- in some cases we likely cannot reproduce, we see the timeout in
'normal' deploys.

The wait_for_backend timeout was added intentionally - to overcome
what was thought to be a race condition between issuing
service mistral-api start, and when the TCP port actually becomes
active.
Having it 'wait for backend' makes the 'workbook create' definitely
succeed.

All that said, this review proposes to drop the timeout to just 1
minute, which we think serves the purpose of allowing mistral-api
to actually startup before workbook create happens.

Change-Id: Ic1b632cc8be662a598be27c7abab1897d076c299